fix(stagewise): restore missing imports and fix formatting in settings refactor
